x/market/types/message_delist_item.go
package types

import (
	sdk "github.com/cosmos/cosmos-sdk/types"
	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
)

var _ sdk.Msg = &MsgDelistItem{}

func NewMsgDelistItem(creator string, id string) *MsgDelistItem {
	return &MsgDelistItem{
		Creator: creator,
		Id:      id,
	}
}

func (msg *MsgDelistItem) Route() string {
	return RouterKey
}

func (msg *MsgDelistItem) Type() string {
	return "DelistItem"
}

func (msg *MsgDelistItem) GetSigners() []sdk.AccAddress {
	creator, err := sdk.AccAddressFromBech32(msg.Creator)
	if err != nil {
		panic(err)
	}
	return []sdk.AccAddress{creator}
}

func (msg *MsgDelistItem) GetSignBytes() []byte {
	bz := ModuleCdc.MustMarshalJSON(msg)
	return sdk.MustSortJSON(bz)
}

func (msg *MsgDelistItem) ValidateBasic() error {
	_, err := sdk.AccAddressFromBech32(msg.Creator)
	if err != nil {
		return sdkerrors.Wrapf(sdkerrors.ErrInvalidAddress, "invalid creator address (%s)", err)
	}
	return nil
}
